Update Readme Due to Prisma Changes
It looks like Prisma updated its migration behavior, which breaks most of the functionality here. Normally upgrading wouldn't be an issue, but the updated migrate flow does not work in Heroku due to permissions. Further, upgrading to 2.13.0 blocks legacy migrate flow. A simple fix is to hold the prisma/cli @ 2.12 until either is fixed. Thanks for the tutorial!